# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "plugin maintenance server api doc",
        "title": "Swagger sync-agent API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "basePath": "/v1",
    "paths": {
        "/synchronization/{platform}/comment": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Synchronization"
                ],
                "summary": "同步 gitee 或 github 平台的 comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "平台：gitee 或 github",
                        "name": "platform",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "同步成功",
                        "schema": {
                            "$ref": "#/definitions/models.BaseResp"
                        }
                    },
                    "404": {
                        "description": "错误返回",
                        "schema": {
                            "$ref": "#/definitions/models.BaseResp"
                        }
                    },
                    "500": {
                        "description": "错误返回",
                        "schema": {
                            "$ref": "#/definitions/models.BaseResp"
                        }
                    }
                }
            }
        },
        "/synchronization/{platform}/issue": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Synchronization"
                ],
                "summary": "同步 gitee 或 github 平台的 issue",
                "parameters": [
                    {
                        "type": "string",
                        "description": "平台：gitee 或 github",
                        "name": "platform",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "同步成功",
                        "schema": {
                            "$ref": "#/definitions/models.BaseResp"
                        }
                    },
                    "404": {
                        "description": "错误返回",
                        "schema": {
                            "$ref": "#/definitions/models.BaseResp"
                        }
                    },
                    "500": {
                        "description": "错误返回",
                        "schema": {
                            "$ref": "#/definitions/models.BaseResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.BaseResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string",
                    "example": "请求成功/失败"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "access-token",
            "in": "header"
        }
    }
}